pyproject.toml
[tool.poetry]
name = "nonebot_plugin_llm_helper"
version = "0.0.1"
description = "一个支持多平台的帮助插件，基于 LLM 解析 Github 仓库或代码。"
license = "GPL3"
authors = ["Lonely-Sails"]
readme = "README.md"
homepage = "https://github.com/Lonely-Sails/nonebot-plugin-llm-helper"
repository = "https://github.com/Lonely-Sails/nonebot-plugin-llm-helper"
documentation = "https://github.com/Lonely-Sails/nonebot-plugin-llm-helper#readme"
keywords = ["nonebot", "nonebot2", "helper", "llm", "ai"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-plugin-localstore = "^0.7.0"
nonebot-plugin-alconna = "^1.8.38"
pydantic = "^2.7.4"
httpx = "^0.28.1"
